+++ packages/ide-extension/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sap-guided-answer-extension
 
+## 1.7.0
+
+### Minor Changes
+
+-   2fab487: Migrate tests from enzyme to react testing library
+
+## 1.6.1
+
+### Patch Changes
+
+-   1f24fdb: Update dependencies
+
 ## 1.6.0
 
 ### Minor Changes